Only copying file if it exists

+++ pippin/merge.py
@@ -46,8 +46,9 @@ class Merger(Task):
                 original = os.path.join(self.lc_fit["lc_output_dir"], f)
                 moved = os.path.join(self.output_dir, f)
                 if not os.path.exists(moved):
-                    self.logger.debug(f"Copying file {f} into output directory")
-                    shutil.move(original, moved)
+                    if os.path.exists(original):
+                        self.logger.debug(f"Copying file {f} into output directory")
+                        shutil.move(original, moved)
 
             # Dick around with folders and names to make it resemble split_and_fit output for salt2mu
             outdir = os.path.join(self.output_dir, self.lc_fit["genversion"])
